+++ domain/src/main/java/com/montfel/domain/usecase/GetNoteByIdUseCase.kt
@@ -3,5 +3,5 @@ package com.montfel.domain.usecase
 import com.montfel.domain.model.Note
 
 interface GetNoteByIdUseCase {
-    suspend operator fun invoke(noteId: Int): Note
+    suspend operator fun invoke(noteId: Long): Note
 }
