+++ presentation/src/main/java/com/montfel/presentation/upsertnote/UpsertNoteEvent.kt
@@ -1,7 +1,7 @@
 package com.montfel.presentation.upsertnote
 
 sealed interface UpsertNoteEvent {
-    data class GetNoteById(val noteId: Int): UpsertNoteEvent
+    data class GetNoteById(val noteId: Long): UpsertNoteEvent
     data object OnUpsertNote: UpsertNoteEvent
     data class OnNoteTitleChange(val title: String) : UpsertNoteEvent
     data class OnNoteDescriptionChange(val description: String) : UpsertNoteEvent
